refactor BYOCRenderer, update unit test, add feaas exteranl
