+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@sitecore-feaas/clientside": "^0.3.10",
+    "@sitecore-feaas/clientside": "^0.3.12",
     "bootstrap": "^5.1.3",
     "font-awesome": "^4.7.0",
     "sass": "^1.52.3",
