+++ packages/sitecore-jss-react/package.json
@@ -61,7 +61,7 @@
     "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@sitecore-feaas/clientside": "^0.3.0",
+    "@sitecore-feaas/clientside": "^0.3.12",
     "@sitecore-jss/sitecore-jss": "21.3.0-canary.45",
     "deep-equal": "^2.1.0",
     "prop-types": "^15.8.1",
