+++ packages/sitecore-jss-react/src/index.ts
@@ -68,7 +68,7 @@ export {
   BYOCRenderer,
   BYOCRenderingParams,
   BYOCRendererProps,
-} from './components/BYOCRenderer';
+} from './components/BYOCComponent';
 export { Link, LinkField, LinkFieldValue, LinkProps, LinkPropTypes } from './components/Link';
 export { File, FileField } from './components/File';
 export { VisitorIdentification } from './components/VisitorIdentification';
